% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaDecideTablesDS.R
\name{limma.decideTablesDS}
\alias{limma.decideTablesDS}
\title{limma.decideTablesDS}
\usage{
limma.decideTablesDS(table.name = NULL)
}
\arguments{
\item{table.name}{the name of the decide table to be returned}
}
\description{
The serverside function called by ds.limma.decideTables
}
\details{
Returns a decide table
}
\author{
DataSHIELD team
}
