% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaTopTableDS.R
\name{limma.topTableDS}
\alias{limma.topTableDS}
\title{limma.topTableDS}
\usage{
limma.topTableDS(table.name = NULL)
}
\arguments{
\item{table.name}{the name of the decide table to be returned}
}
\description{
The serverside function called by ds.limma.topTable
}
\details{
Returns a decide table
}
\author{
DataSHIELD team
}
