% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaDecideTablesDS.R
\name{limmaDecideTablesDS}
\alias{limmaDecideTablesDS}
\title{limmaDecideTablesDS}
\usage{
limmaDecideTablesDS(table.name = NULL)
}
\arguments{
\item{table.name}{the name of the decide table to be returned}
}
\description{
The serverside function called by ds.limmaDecideTables
}
\details{
Returns a decide table
}
\author{
DataSHIELD team
}
