% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaTopTableDS.R
\name{limmaTopTableDS}
\alias{limmaTopTableDS}
\title{limmaTopTableDS}
\usage{
limmaTopTableDS(table.name = NULL)
}
\arguments{
\item{table.name}{the name of the decide table to be returned}
}
\description{
The serverside function called by ds.limmaTopTable
}
\details{
Returns a decide table
}
\author{
DataSHIELD team
}
